[ACE6_TAO2] Fixed warnings and updated bison-generated code for ACE svc.conf

(cherry picked from commit 9f734b46e79a3c518a1d9d2d2727495cc68f17a5)

# Conflicts:
#	.github/workflows/linux.yml
#	ACE/ace/Asynch_IO_Impl.inl
#	ACE/ace/Svc_Conf_Token_Table.h
#	ACE/ace/Svc_Conf_y.cpp
#	TAO/tao/IIOP_Endpoint.cpp
#	TAO/tao/LF_CH_Event.cpp
#	TAO/tao/Object.cpp
#	TAO/tao/Storable_FlatFileStream.cpp

+++ ACE/ace/ICMP_Socket.cpp
@@ -172,7 +172,7 @@ ACE_ICMP_Socket::calculate_checksum (unsigned short * paddress,
   // add back carry outs from top 16 bits to low 16 bits
   sum = (sum >> 16) + (sum & 0xffff); // add hi 16 to low 16
   sum += (sum >> 16);                 // add carry
-  answer = ~sum;                      // truncate to 16 bits
+  answer = static_cast<unsigned short> (~sum); // truncate to 16 bits
 
   return (answer);
 }
